+++ src/dockerfiles/eslint__eslint-11407.Dockerfile
@@ -52,6 +52,7 @@ ENV DBUS_SESSION_BUS_ADDRESS="unix:path=/run/dbus/system_bus_socket"
 RUN dbus-daemon --system --fork
 
 ENV PUPPETEER_SKIP_CHROMIUM_DOWNLOAD=true
+ENV PUPPETEER_SKIP_DOWNLOAD=true
 ENV OPENSSL_CONF /etc/ssl
 
 RUN useradd -m chromeuser
@@ -88,33 +89,31 @@ python2 -V
 EOF_31553638dfda
 
 
-RUN <<EOF_83f882f70c9e
+RUN <<EOF_23af5fa81cf7
 #!/bin/bash
 set -euxo pipefail
 git clone -o origin https://github.com/eslint/eslint /testbed
-chmod -R 777 /testbed
 cd /testbed
 git reset --hard af9688b7c4f6a3afe1b0ca5ba2f475c545e0309b
 git remote remove origin
-TARGET_EPOCH=$(git show -s --format=%ct af9688b7c4f6a3afe1b0ca5ba2f475c545e0309b)
-git tag -l | while read tag; do TAG_COMMIT=$(git rev-list -n 1 "$tag"); TAG_EPOCH=$(git show -s --format=%ct "$TAG_COMMIT"); if [ "$TAG_EPOCH" -gt "$TARGET_EPOCH" ]; then git tag -d "$tag"; fi; done
-git branch -D $(git branch | grep -v "^\*") 2>/dev/null || true
+git branch | grep -v '^\*' | xargs -r git branch -D || true
+git tag -l | xargs -r git tag -d
 git reflog expire --expire=now --all
+git gc --prune=now --aggressive
+TARGET_EPOCH=$(git show -s --format=%ct af9688b7c4f6a3afe1b0ca5ba2f475c545e0309b)
+AFTER_EPOCH=$((TARGET_EPOCH + 1))
+AFTER_TIMESTAMP=$(date -u -d "@$AFTER_EPOCH" "+%Y-%m-%d %H:%M:%S")
+COMMIT_COUNT=$(git log --oneline --all --since="$AFTER_TIMESTAMP" | wc -l)
+[ "$COMMIT_COUNT" -eq 0 ] || exit 1
 cd - || true
+chmod -R 777 /testbed
 cd /testbed
 git clean -fdxq
 source $NVM_DIR/nvm.sh
 npm install
-EOF_83f882f70c9e
+EOF_23af5fa81cf7
 
 
-RUN <<EOF_9c175e27b1a0
-#!/bin/bash
-set -euxo pipefail
-mkdir -p /swebench/image_assets
-mkdir -p /swebench/image_assets/problem_statement
-curl -fsSL -o '/swebench/image_assets/problem_statement/51068445-e0c78b80-15d2-11e9-865d-17659d123e3e.png' 'https://user-images.githubusercontent.com/15365418/51068445-e0c78b80-15d2-11e9-865d-17659d123e3e.png' || true
-EOF_9c175e27b1a0
-
+COPY src/image_assets/eslint__eslint-11407/ /swebench/image_assets/
 
 WORKDIR /testbed
