+++ src/dockerfiles/openlayers__openlayers-10340.Dockerfile
@@ -52,6 +52,7 @@ ENV DBUS_SESSION_BUS_ADDRESS="unix:path=/run/dbus/system_bus_socket"
 RUN dbus-daemon --system --fork
 
 ENV PUPPETEER_SKIP_CHROMIUM_DOWNLOAD=true
+ENV PUPPETEER_SKIP_DOWNLOAD=true
 ENV OPENSSL_CONF /etc/ssl
 
 RUN useradd -m chromeuser
@@ -91,36 +92,34 @@ python2 -V
 EOF_55f960f4ac15
 
 
-RUN <<EOF_8344fe65086a
+RUN <<EOF_4c90ab67d394
 #!/bin/bash
 set -euxo pipefail
 git clone -o origin https://github.com/openlayers/openlayers /testbed
-chmod -R 777 /testbed
 cd /testbed
 git reset --hard e361391503ada8a9efdce2f3aa75e29d0351ab95
 git remote remove origin
-TARGET_EPOCH=$(git show -s --format=%ct e361391503ada8a9efdce2f3aa75e29d0351ab95)
-git tag -l | while read tag; do TAG_COMMIT=$(git rev-list -n 1 "$tag"); TAG_EPOCH=$(git show -s --format=%ct "$TAG_COMMIT"); if [ "$TAG_EPOCH" -gt "$TARGET_EPOCH" ]; then git tag -d "$tag"; fi; done
-git branch -D $(git branch | grep -v "^\*") 2>/dev/null || true
+git branch | grep -v '^\*' | xargs -r git branch -D || true
+git tag -l | xargs -r git tag -d
 git reflog expire --expire=now --all
+git gc --prune=now --aggressive
+TARGET_EPOCH=$(git show -s --format=%ct e361391503ada8a9efdce2f3aa75e29d0351ab95)
+AFTER_EPOCH=$((TARGET_EPOCH + 1))
+AFTER_TIMESTAMP=$(date -u -d "@$AFTER_EPOCH" "+%Y-%m-%d %H:%M:%S")
+COMMIT_COUNT=$(git log --oneline --all --since="$AFTER_TIMESTAMP" | wc -l)
+[ "$COMMIT_COUNT" -eq 0 ] || exit 1
 cd - || true
+chmod -R 777 /testbed
 cd /testbed
 git clean -fdxq
 source $NVM_DIR/nvm.sh
 npm install
 sed -i "s|process.env.CHROME_BIN = require('puppeteer').executablePath();|process.env.CHROME_BIN = '/usr/bin/google-chrome-stable';|" test/karma.config.js
-EOF_8344fe65086a
+npm install karma-json-reporter@1.2.1 --no-save --legacy-peer-deps
+sed -i "s/reporters: \['dots', 'coverage-istanbul'\]/reporters: ['json'],\n        jsonReporter: { stdout: true }/" test/karma.config.js ; sed -i "s/reporters: \['dots'\]/reporters: ['json'],\n        jsonReporter: { stdout: true }/" test/karma.config.js ; sed -i "s/reporters: \['progress'\]/reporters: ['json'],\n        jsonReporter: { stdout: true }/" test/karma.config.js
+EOF_4c90ab67d394
 
 
-RUN <<EOF_1ca880b5835e
-#!/bin/bash
-set -euxo pipefail
-mkdir -p /swebench/image_assets
-mkdir -p /swebench/image_assets/problem_statement
-curl -fsSL -o '/swebench/image_assets/problem_statement/69544237-373d7e80-0f87-11ea-9303-b5be8a4c1ed4.png' 'https://user-images.githubusercontent.com/49240900/69544237-373d7e80-0f87-11ea-9303-b5be8a4c1ed4.png' || true
-mkdir -p /swebench/image_assets/problem_statement
-curl -fsSL -o '/swebench/image_assets/problem_statement/69544386-85528200-0f87-11ea-9f83-02ed50f36724.png' 'https://user-images.githubusercontent.com/49240900/69544386-85528200-0f87-11ea-9f83-02ed50f36724.png' || true
-EOF_1ca880b5835e
-
+COPY src/image_assets/openlayers__openlayers-10340/ /swebench/image_assets/
 
 WORKDIR /testbed
